feat: consolidate common data processing actions into new class
