+++ timecopilot/models/foundation/flowstate.py
@@ -3,12 +3,11 @@ from contextlib import contextmanager
 import numpy as np
 import pandas as pd
 import torch
-from gluonts.transform import LastValueImputation
 from tqdm import tqdm
 from tsfm_public import FlowStateForPrediction
 from tsfm_public.models.flowstate.utils.utils import get_fixed_factor
 
-from ..utils.forecaster import Forecaster, QuantileConverter
+from ..utils.forecaster import Forecaster, QuantileConverter, _DataProcessor
 from .utils import TimeSeriesDataset
 
 
@@ -99,49 +98,6 @@ class FlowState(Forecaster):
             del model
             torch.cuda.empty_cache()
 
-    def _left_pad_and_stack_1D(self, tensors: list[torch.Tensor]) -> torch.Tensor:
-        max_len = max(len(c) for c in tensors)
-        padded = []
-        for c in tensors:
-            assert isinstance(c, torch.Tensor)
-            assert c.ndim == 1
-            padding = torch.full(
-                size=(max_len - len(c),),
-                fill_value=torch.nan,
-                device=c.device,
-                dtype=c.dtype,
-            )
-            padded.append(torch.concat((padding, c), dim=-1))
-        return torch.stack(padded)
-
-    def _prepare_and_validate_context(
-        self,
-        context: list[torch.Tensor] | torch.Tensor,
-    ) -> torch.Tensor:
-        if isinstance(context, list):
-            context = self._left_pad_and_stack_1D(context)
-        assert isinstance(context, torch.Tensor)
-        if context.ndim == 1:
-            context = context.unsqueeze(0)
-        assert context.ndim == 2
-        return context
-
-    def _maybe_impute_missing(self, batch: torch.Tensor) -> torch.Tensor:
-        if torch.isnan(batch).any():
-            batch = batch.float().numpy()
-            imputed_rows = []
-            for i in range(batch.shape[0]):
-                row = batch[i]
-                imputed_row = LastValueImputation()(row)
-                imputed_rows.append(imputed_row)
-            batch = np.vstack(imputed_rows)
-            batch = torch.tensor(
-                batch,
-                dtype=self.dtype,
-                device=self.device,
-            )
-        return batch
-
     def _predict_batch(
         self,
         model: FlowStateForPrediction,
@@ -151,10 +107,11 @@ class FlowState(Forecaster):
         supported_quantiles: list[float],
         scale_factor: float,
     ) -> tuple[np.ndarray, np.ndarray | None]:
-        context = self._prepare_and_validate_context(batch)
+        data_processor = _DataProcessor(dtype=self.dtype, device=self.device)
+        context = data_processor._prepare_and_validate_context(batch)
         if context.shape[1] > self.context_length:
             context = context[..., -self.context_length :]
-        context = self._maybe_impute_missing(context)
+        context = data_processor._maybe_impute_missing(context)
         # context is (batch, context_length)
         # then we convert it to (context_length, batch, 1)
         context = context.unsqueeze(-1).transpose(0, 1)
