+++ timecopilot/models/foundation/patchtst_fm.py
@@ -3,11 +3,10 @@ from contextlib import contextmanager
 import numpy as np
 import pandas as pd
 import torch
-from gluonts.transform import LastValueImputation
 from tqdm import tqdm
 from tsfm_public import PatchTSTFMForPrediction
 
-from ..utils.forecaster import Forecaster, QuantileConverter
+from ..utils.forecaster import Forecaster, QuantileConverter, _DataProcessor
 from .utils import TimeSeriesDataset
 
 # default to the median quantile
@@ -104,51 +103,6 @@ class PatchTSTFM(Forecaster):
             elif self.device.startswith("mps"):
                 torch.mps.empty_cache()
 
-    def _left_pad_and_stack_1D(self, tensors: list[torch.Tensor]) -> torch.Tensor:
-        max_len = max(len(c) for c in tensors)
-        padded = []
-        for c in tensors:
-            assert isinstance(c, torch.Tensor)
-            assert c.ndim == 1
-            padding = torch.full(
-                size=(max_len - len(c),),
-                fill_value=torch.nan,
-                device=c.device,
-                dtype=c.dtype,
-            )
-            padded.append(torch.concat((padding, c), dim=-1))
-        return torch.stack(padded)
-
-    def _prepare_and_validate_context(
-        self,
-        context: list[torch.Tensor] | torch.Tensor,
-    ) -> torch.Tensor:
-        if isinstance(context, list):
-            context = self._left_pad_and_stack_1D(context)
-        assert isinstance(context, torch.Tensor)
-        if context.ndim == 1:
-            context = context.unsqueeze(0)
-        assert context.ndim == 2
-        return context
-
-    def _maybe_impute_missing(
-        self, batch: torch.Tensor, dtype=torch.float32
-    ) -> torch.Tensor:
-        if torch.isnan(batch).any():
-            batch = batch.to(dtype=dtype).numpy(force=True)
-            imputed_rows = []
-            for i in range(batch.shape[0]):
-                row = batch[i]
-                imputed_row = LastValueImputation()(row)
-                imputed_rows.append(imputed_row)
-            batch = np.vstack(imputed_rows)
-            batch = torch.tensor(
-                batch,
-                dtype=self.dtype,
-                device=self.device,
-            )
-        return batch
-
     def _predict_batch(
         self,
         model: PatchTSTFMForPrediction,
@@ -157,10 +111,11 @@ class PatchTSTFM(Forecaster):
         quantiles: list[float] | None,
         # scale_factor: float,
     ) -> tuple[np.ndarray, np.ndarray | None]:
-        context = self._prepare_and_validate_context(batch)
+        data_processor = _DataProcessor(dtype=self.dtype, device=self.device)
+        context = data_processor._prepare_and_validate_context(batch)
         if context.shape[1] > self.context_length:
             context = context[..., -self.context_length :]
-        context = self._maybe_impute_missing(context)
+        context = data_processor._maybe_impute_missing(context)
         # context is (batch, context_length)
 
         # input data is grouped by id
@@ -180,12 +135,11 @@ class PatchTSTFM(Forecaster):
         # may not be the ideal solution, but this should be more adaptable
         # when quantiles can vary.
         # there is no guarantee that 0.5 will be in the list of quantiles.
-        mean = fcst.mean(dim=-1).squeeze() if fcst.ndim >= 3 else fcst.squeeze()
+        fcst_mean = fcst.mean(dim=-1).squeeze() if fcst.ndim >= 3 else fcst.squeeze()
         # fcst_mean = fcst[..., quantile_levels.index(0.5)].squeeze()
-        fcst_mean = mean
-        fcst_mean_np = fcst_mean.detach().numpy(force=True)
+        fcst_mean_np = fcst_mean.detach().cpu().numpy()
         fcst_quantiles_np = (
-            fcst.detach().numpy(force=True) if quantiles is not None else None
+            fcst.detach().cpu().numpy() if quantiles is not None else None
         )
         return fcst_mean_np, fcst_quantiles_np
